pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentflow"
version = "0.1.0"
description = "AgentFlow planning and execution utilities with Codex CLI integration."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0",
    "pydantic>=2.8",
    "PyYAML>=6.0",
    "Flask>=3.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-dotenv>=0.5",
    "pytest-mock>=3.12"
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning"
]

[project.scripts]
agentflow = "agentflow.cli:main"
